Added a Modelica interface (OMSimulatorModelica)
A workaround is needed in order to execute the accompanying tests.
In the top level CMakeLists.txt the line
add_definitions(-DBTH_DEACTIVATE_INITIAL_UNKNOWNS)
must be enabled. Without it, the cs_BouncingBall.mos test in
OMSimulatorModelica fails with a segmentation fault. Strangely,
executing the stimulated example model in OMEdit instead of from the
command line works without the workaround. Maybe different shared
libraries are loaded for the two cases. Needs investigation.